+++ src/app/(routes)/(groups)/games/[slug]/_helpers/stats.ts
