+++ src/app/(routes)/(groups)/games/[slug]/page.tsx
@@ -1,14 +1,14 @@
 import { H1 } from "@/components/headings";
 import { getAllGames } from "../../../../../../prisma/lib/games";
 import { getAllSessionsByGame } from "../../../../../../prisma/lib/admin"; // Import getAllSessions
-import Mariokart from "./_components/mariokart";
-import CallOfDuty from "./_components/callofduty";
-import RocketLeague from "./_components/rocketleague";
-import Speedrunners from "./_components/speedrunners";
-import LethalCompany from "./_components/lethalcompany";
+import Mariokart from "./_components/games/mariokart";
+import CallOfDuty from "./_components/games/callofduty";
+import RocketLeague from "./_components/games/rocketleague";
+import Speedrunners from "./_components/games/speedrunners";
+import LethalCompany from "./_components/games/lethalcompany";
 // import GolfWithFriends from "./_components/golfwithfriends";
 import { gameImages, GamesEnum } from "@/lib/constants";
-import { TimelineChart } from "./_components/timeline-chart";
+import { TimelineChart } from "./_components/timeline/timeline-chart";
 import { Separator } from "@/components/ui/separator";
 
 // ? Force non specified routes to return 404
